refactor(analytics, fiam): use BoM 32.5.0 and drop ktx
