+++ inappmessaging/app/build.gradle.kts
@@ -52,7 +52,7 @@ dependencies {
     implementation("androidx.multidex:multidex:2.0.1")
 
     // Import the Firebase BoM (see: https://firebase.google.com/docs/android/learn-more#bom)
-    implementation(platform("com.google.firebase:firebase-bom:32.4.1"))
+    implementation(platform("com.google.firebase:firebase-bom:32.5.0"))
 
     // FIAM
     implementation("com.google.firebase:firebase-inappmessaging-display")
